+++ lib/DidWebDriver.js
@@ -3,16 +3,15 @@
  */
 import * as didIo from '@digitalbazaar/did-io';
 import {
-  assertHttpsUrl,
-  didToUrl,
-  urlToDid
-} from './helpers.js';
-import {
   contextsBySuite,
   defaultFetchOptions,
   DID_CONTEXT_URL,
   ED25519_KEY_2018_CONTEXT_URL
 } from './constants.js';
+import {
+  didToUrl,
+  urlToDid
+} from './helpers.js';
 import {
   Ed25519VerificationKey2020
 } from '@digitalbazaar/ed25519-verification-key-2020';
@@ -69,7 +68,6 @@ export class DidWebDriver {
    *   KMS).
    */
   async generate({url, seed} = {}) {
-    assertHttpsUrl(url);
     // Public/private key pair of the main did:key signing/verification key
     const verificationKeyPair = await this.verificationSuite.generate({seed});
