+++ lib/constants.js
@@ -31,3 +31,4 @@ export const defaultFetchOptions = {
   timeout: 5000
 };
 
+export const didPrefix = 'did:web:';
